pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xaiops-graphrag"
version = "0.1.0"
description = "GraphRAG application using Neo4j and LangGraph"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "tiktoken>=0.7.0",
    "httpx>=0.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]
